CoordinateCalculationクラスが様々な処理を請け負っていて実状と合っていなかったため、DiagramEditorDocumentクラスとTimeCalclulationクラスに処理を委譲した。
