--- DiagramEditor/View/DiagramView/TimeCalculation.swift
@@ -1,8 +1,0 @@
-//
-//  TimeCalculation.swift
-//  DiagramEditor
-//
-//  Created by 丹羽雄一朗 on 2023/12/18.
-//
-
-import Foundation
